+++ common/src/main/java/whocraft/tardis_refined/client/model/blockentity/shell/ShellModelCollection.java
@@ -115,7 +115,7 @@ public class ShellModelCollection {
 
     private void validateModels() {
         for (ResourceLocation resourceLocation : ShellTheme.SHELL_THEME_DEFERRED_REGISTRY.keySet()) {
-            if(SHELL_MODELS.containsKey(resourceLocation)){
+            if(!SHELL_MODELS.containsKey(resourceLocation)){
                 TardisRefined.LOGGER.info("There was no model setup for shell theme {}", resourceLocation);
             }
         }