+++ common/src/main/java/whocraft/tardis_refined/common/tardis/TardisDesktops.java
@@ -80,6 +80,7 @@ public class TardisDesktops {
         addDefaultDesktop(new DesktopTheme("violet_eye", "desktop/violet_eye"));
         addDefaultDesktop(new DesktopTheme("watchface", "desktop/watchface"));
         addDefaultDesktop(new DesktopTheme("classic", "desktop/classic"));
+        addDefaultDesktop(new DesktopTheme("terraformed", "desktop/terraformed_useable"));
 
         return DEFAULT_DESKTOPS;
     }